refactor(common)!: move expression package to org.apache.hudi.common.expression
Relocates the engine-agnostic predicate/expression AST (Expression,
Predicates, Literal, NameReference, bind visitors, StructLike/ArrayData)
from org.apache.hudi.expression to org.apache.hudi.common.expression.
The package has no Hudi domain awareness - its only dependencies are
the internal type system and Option - and its consumers span the read
path, engine bridges and metastore sync tools alike, making it shared
substrate (common) rather than domain machinery (core). Keeps the
dependency layering acyclic for the future module split:
types -> expressions -> domain -> engines.

BREAKING CHANGE: source/binary break for external code constructing
Hudi expressions programmatically, e.g. callers of
HoodieTableMetadata.getPartitionPathWithPathPrefixUsingFilterExpression.
Fails loudly; fixed by an import change. Expression classes are never
named in configs and never loaded reflectively.
