+++ hudi-common/src/main/java/org/apache/hudi/common/expression/Predicate.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.expression;
+package org.apache.hudi.common.expression;
 
 import org.apache.hudi.internal.schema.Type;
 import org.apache.hudi.internal.schema.Types;
