+++ hudi-common/src/main/java/org/apache/hudi/core/read/BaseHoodieTableFileIndex.java
@@ -22,7 +22,7 @@ import org.apache.hudi.common.config.HoodieMemoryConfig;
 import org.apache.hudi.common.config.HoodieMetadataConfig;
 import org.apache.hudi.common.config.TypedProperties;
 import org.apache.hudi.common.engine.HoodieEngineContext;
-import org.apache.hudi.expression.Expression;
+import org.apache.hudi.common.expression.Expression;
 import org.apache.hudi.common.fs.FSUtils;
 import org.apache.hudi.common.model.BaseFile;
 import org.apache.hudi.common.model.FileSlice;
