+++ hudi-common/src/main/java/org/apache/hudi/metadata/AbstractHoodieTableMetadata.java
@@ -19,9 +19,9 @@
 package org.apache.hudi.metadata;
 
 import org.apache.hudi.common.engine.HoodieEngineContext;
+import org.apache.hudi.common.expression.ArrayData;
 import org.apache.hudi.common.util.PartitionPathEncodeUtils;
 import org.apache.hudi.common.util.StringUtils;
-import org.apache.hudi.expression.ArrayData;
 import org.apache.hudi.internal.schema.Type;
 import org.apache.hudi.internal.schema.Types;
 import org.apache.hudi.storage.HoodieStorage;
