+++ hudi-common/src/test/java/org/apache/hudi/common/table/log/TestHoodieMergedLogRecordReader.java
@@ -19,11 +19,11 @@
 
 package org.apache.hudi.common.table.log;
 
+import org.apache.hudi.common.expression.Expression;
+import org.apache.hudi.common.expression.Literal;
+import org.apache.hudi.common.expression.Predicate;
+import org.apache.hudi.common.expression.Predicates;
 import org.apache.hudi.common.util.Option;
-import org.apache.hudi.expression.Expression;
-import org.apache.hudi.expression.Literal;
-import org.apache.hudi.expression.Predicate;
-import org.apache.hudi.expression.Predicates;
 
 import org.junit.jupiter.api.Test;
