+++ hudi-common/src/test/java/org/apache/hudi/metadata/TestHoodieBackedTableMetadataBuildPredicate.java
@@ -18,8 +18,8 @@
 
 package org.apache.hudi.metadata;
 
-import org.apache.hudi.expression.Expression;
-import org.apache.hudi.expression.Predicate;
+import org.apache.hudi.common.expression.Expression;
+import org.apache.hudi.common.expression.Predicate;
 
 import org.junit.jupiter.api.Test;
