+++ hudi-spark-datasource/hudi-spark-common/src/main/scala/org/apache/hudi/SparkFilterHelper.scala
@@ -19,7 +19,7 @@
 
 package org.apache.hudi
 
-import org.apache.hudi.expression.{Expression, Literal, NameReference, Predicates}
+import org.apache.hudi.common.expression.{Expression, Literal, NameReference, Predicates}
 import org.apache.hudi.internal.schema.{Type, Types}
 
 import org.apache.spark.sql.catalyst.util.DateTimeUtils
